requirements.txt
timm==0.6.7
deepspeed==0.9.3
data
einops==0.6.1
ftfy==6.1.1
iopath==0.1.10
ipdb==0.13.13
numpy==1.24.3
peft==0.3.0
Pillow==9.5.0
PyYAML==6.0
regex==2022.10.31
pytorchvideo
fvcore
decord==0.6.0
tqdm==4.64.1
setuptools==65.5.1
transformers==4.35.0
bigmodelvis
nltk
tensorboard
cython
plyfile
trimesh
sentencepiece
albumentations==0.4.3
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.0.8
omegaconf==2.0.0
test-tube>=0.7.5
streamlit>=0.73.1
more-itertools>=8.0.0
colorama
lpips